requirements.txt
alembic==1.15.2
fastapi==0.115.12
geoalchemy2==0.17.1
importlib-metadata==8.0.0
importlib-resources==6.4.0
jaraco.text==3.12.1
line-bot-sdk==3.16.3
loguru==0.7.3
marshmallow==4.0.0
mysqlclient==2.2.7
ngrok==1.4.0
pip-chill==1.0.3
platformdirs==4.2.2
pydantic-settings==2.9.1
pymongo==4.12.1
pymysql==1.1.1
redis==5.2.1
shapely==2.1.0
tomli==2.0.1
uvicorn==0.34.1
